исправлена ошибка в тестах
